requirements.txt
pip~=24.2
pillow~=10.4.0
redis~=5.1.1
Django~=5.1.1
asgiref~=3.8.1
sqlparse~=0.5.1
psycopg2~=2.9.9
psycopg2-binary~=2.9.9
djangorestframework~=3.15.2
djangorestframework-simplejwt~=5.3.1
django-environ~=0.11.2